backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
cryptography>=41.0.0
greenlet>=3.0.0
